Refactors idle manager to shortcut call check
